+++ Vienna/Export.m
@@ -60,7 +60,8 @@
 	// mysterious.
 	if (![[NSFileManager defaultManager] createFileAtPath:pathToFile contents:nil attributes:nil])
 	{
-		NSBeginCriticalAlertSheet(NSLocalizedString(@"Cannot open export file message", nil),
+#warning 64BIT: Check formatting arguments
+		NSBeginCriticalAlertSheet (NSLocalizedString(@"Cannot open export file message", nil),
 								  NSLocalizedString(@"OK", nil),
 								  nil,
 								  nil, [NSApp mainWindow], self,
@@ -124,6 +125,7 @@
 	BOOL isSorted;
 	
 	NSString * folderPath = [db folderPathName:[folder itemId]];
+#warning 64BIT: Check formatting arguments
 	NSString * progressText = [NSString stringWithFormat:NSLocalizedString(@"Retrieving messages from %@", nil), folderPath];
 	[self updateProgressText:progressText];
 	
@@ -158,6 +160,7 @@
 	NSAutoreleasePool * pool;
     pool = [[NSAutoreleasePool alloc] init];
 	
+#warning 64BIT: Check formatting arguments
 	NSString * progressText = [NSString stringWithFormat:NSLocalizedString(@"Opening '%@'", nil), exportFilename];
 	[self performSelectorOnMainThread:@selector(updateProgressText:) withObject:progressText waitUntilDone:YES];
 	
@@ -197,11 +200,13 @@
 		NSString * folderPath = [db folderPathName:[theMessage folderId]];
 		NSCalendarDate * theDate = [[theMessage date] dateWithCalendarFormat:nil timeZone:nil];
 		NSString * dateString = [theDate descriptionWithCalendarFormat:@"%d%b%y %H:%M"];
+#warning 64BIT: Check formatting arguments
 		NSString * commentString = [theMessage comment] > 0 ? [NSString stringWithFormat:@" c%d", [theMessage comment]] : @"";
 		
 		// Only report progress on change of folder
 		if ([theMessage folderId] != lastFolderId)
 		{
+#warning 64BIT: Check formatting arguments
 			NSString * progressText = [NSString stringWithFormat:NSLocalizedString(@"Exporting from %@", nil), folderPath];
 			[self performSelectorOnMainThread:@selector(updateProgressText:) withObject:progressText waitUntilDone:YES];
 			lastFolderId = [theMessage folderId];
@@ -209,7 +214,7 @@
 		
 		// Call the main thread to retrieve the text of a specific message
 		[self performSelectorOnMainThread:@selector(loadMessageText:) withObject:theMessage waitUntilDone:YES];
-		int size = [messageText length];
+		NSInteger size = [messageText length];
 		
 		NSMutableString * msgStatus = [NSMutableString stringWithString:@"!MF:"];
 		if ([theMessage isPriority])
@@ -221,6 +226,7 @@
 		if ([theMessage isFlagged])
 			[msgStatus appendString:@"M"];
 		
+#warning 64BIT: Check formatting arguments
 		NSString * msgText = [NSString stringWithFormat:@"%@\n>>>%@ %d %@(%d)%@%@\n%@\n",
 			msgStatus,
 			folderPath,
@@ -293,7 +299,7 @@
 /* exportSavePanelDidEnd
  * Called when the user completes the Export save panel
  */
--(void)exportSavePanelDidEnd:(NSSavePanel *)panel returnCode:(int)returnCode contextInfo:(void *)contextInfo
+-(void)exportSavePanelDidEnd:(NSSavePanel *)panel returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo
 {
 	if (returnCode == NSOKButton)
 	{
@@ -326,7 +332,7 @@
 {
 	XMLParser * newTree = [[XMLParser alloc] initWithEmptyTree];
 	XMLParser * opmlTree = [newTree addTree:@"opml" withAttributes:[NSDictionary dictionaryWithObject:@"1.0" forKey:@"version"]];
-	int countExported = 0;
+	NSInteger countExported = 0;
 
 	// Create the header section
 	XMLParser * headTree = [opmlTree addTree:@"head"];
@@ -362,7 +368,8 @@
 	// Now write the complete XML to the file
 	if (![[NSFileManager defaultManager] createFileAtPath:exportFileName contents:nil attributes:nil])
 	{
-		NSBeginCriticalAlertSheet(NSLocalizedString(@"Cannot open export file message", nil),
+#warning 64BIT: Check formatting arguments
+		NSBeginCriticalAlertSheet (NSLocalizedString(@"Cannot open export file message", nil),
 								  NSLocalizedString(@"OK", nil),
 								  nil,
 								  nil, [NSApp mainWindow], self,
@@ -382,6 +389,7 @@
 	}
 
 	// Announce how many we successfully imported
+#warning 64BIT: Check formatting arguments
 	NSString * successString = [NSString stringWithFormat:NSLocalizedString(@"%d subscriptions successfully exported", nil), countExported];
 	NSRunAlertPanel(NSLocalizedString(@"RSS Subscription Export Title", nil), successString, NSLocalizedString(@"OK", nil), nil, nil);
